+++ R/diameter.R
@@ -115,31 +115,3 @@ fm_diameter.fm_mesh_2d <- function(x, ...) {
 fm_diameter.fm_mesh_1d <- function(x, ...) {
   diff(x[["interval"]])
 }
-
-#' @rdname fm_diameter
-#' @export
-#' @method fm_diameter inla.mesh.1d
-fm_diameter.inla.mesh.1d <- function(x, ...) {
-  fm_diameter(fm_as_fm(x), ...)
-}
-
-#' @rdname fm_diameter
-#' @export
-#' @method fm_diameter inla.mesh.segment
-fm_diameter.inla.mesh.segment <- function(x, ...) {
-  fm_diameter(fm_as_fm(x), ...)
-}
-
-#' @rdname fm_diameter
-#' @export
-#' @method fm_diameter inla.mesh.lattice
-fm_diameter.inla.mesh.lattice <- function(x, ...) {
-  fm_diameter(fm_as_fm(x), ...)
-}
-
-#' @rdname fm_diameter
-#' @export
-#' @method fm_diameter inla.mesh
-fm_diameter.inla.mesh <- function(x, ...) {
-  fm_diameter(fm_as_fm(x), ...)
-}